Audit code to ensure printf style templates are compile time constants.
When the format string is not, it could be imfluenced by error messages
or user input, and could possibly include illegal format specifiers
(such as %p). While not as dangerous as in C/C++, this can cause
exceptions to be thrown, which is especially bad, since the reason we're
printf'ing is typically because we're in the middle of error handling.

As part of this, I've audited usages of:
- f.f.util.Assert.hardAssert
- f.f.util.Assert.fail
- f.f.util.Logger.warn
- f.f.util.Logger.debug
- java.lang.String.format