Add shutdown logic...
... to the Datastore, AsyncQueue, FirestoreChannel. Without this, we get
a potential resource leak when Firestore shuts down. Currently, it only
shuts down in tests, so this doesn't impact production, however, with
the (soon-to-come) AndroidManagedChannel change, the integration tests
will fail, as the Android OS itself tracks resources associated with
AndroidManagedChannel and will terminate if they're not properly cleaned
up.