Switch .getClass() to instanceof in some .equals() methods
This is for cases where we're dealing with a non-leaf node in the
inheritance hierarchy. (Leaf nodes in the heirarchy can just mark the
entire class as final.) In order to preserve symmetry of .equals(), also
mark the .equals() methods themselves as final.

In this case, "non-leaf" nodes also counts effectively final classes
which are mocked. (i.e. ViewSnapshot.)

https://errorprone.info/bugpattern/EqualsGetClass