Cleanup a number of warnings in the code. (#137)
* Eliminate 'badimport' warning

It was already reasonably obvious what this is, so the warning is mostly
spurious. But it's easier and less boilerplate to just add the extra
qualifier than to suppress the warning.

https://errorprone.info/bugpattern/BadImport

* Suppress EqualsGetClass warnings for public classes

These are effectively final and are documented as not being final only
to allow mocking within tests.

https://errorprone.info/bugpattern/EqualsGetClass

* Mark a number of classes final

Mostly to avoid a warning, though it's generally recommended to do this
anyways. (Note that this makes mocking impossible; if any of these
classes need to be mocked, we might be better of just suppressing the
warning or using instanceof.)

https://errorprone.info/bugpattern/EqualsGetClass

* Switch .getClass() to instanceof in some .equals() methods

This is for cases where we're dealing with a non-leaf node in the
inheritance hierarchy. (Leaf nodes in the heirarchy can just mark the
entire class as final.) In order to preserve symmetry of .equals(), also
mark the .equals() methods themselves as final.

In this case, "non-leaf" nodes also counts effectively final classes
which are mocked. (i.e. ViewSnapshot.)

https://errorprone.info/bugpattern/EqualsGetClass

* Provide toString on FieldMask

Since it's used by PatchMutation

https://errorprone.info/bugpattern/ObjectToString

* Fix a 'missing override' warning

https://errorprone.info/bugpattern/MissingOverride

* Rework FieldMask to use a Set of FieldPaths rather than a Collection.

FieldMask uses .equals() on the container, and that's not well defined
for a Collection. The fix is to switch to either a Set or a List. Set
seems more appropriate.

https://errorprone.info/bugpattern/UndefinedEquals

* Add TestUtil.fieldMask(String...) to simplify creating FieldMasks.