Fix multi-process concurrency issues.
On some API levels it is impossible to open a connection to the database
when an "exclusive" transaction is in progress. Although on some others
it is possible to open it but not possible to start a transaction. This
commit includes a change that makes behavior consistent across versions
all the way down to api=16.

Additionally fix data races in instrumentation tests.