[datatransport] Implement EventStore backed by SQLite. (#284)
* [datatransport] Implement EventStore backed by SQLite.

The database contains 3 tables:

* `events`: contains individual events
* `event_metadata`: contains many-to-one metadata key-value pairs
  associated with events
* `backends`: contains metadata about known/registered backends

* Remove android test dependencies.

* Update sqlite interface and introduce a locking mechanism.

The locking mechanism has mutex-like behavior and works across process
boundaries. Synchronization is done with sqlite exclusive locking via
"BEGIN EXCLUSIVE;...;COMMIT;".

* Fix multi-process concurrency issues.

On some API levels it is impossible to open a connection to the database
when an "exclusive" transaction is in progress. Although on some others
it is possible to open it but not possible to start a transaction. This
commit includes a change that makes behavior consistent across versions
all the way down to api=16.

Additionally fix data races in instrumentation tests.

* Address review comments.

* More review comments.

* Address review comments.