Adds the concept of TransportContext. (#335)
TransportContext generalizes the grouping of events that was previously
done by "backend". This is not sufficient as we'd want to start grouping
events by additional dimensions like: priority, api_key, etc.

The change is mostly forward-looking as it will allow us to simplify
storage upgrades as we implement more dimensions, priority being the
main one. And it boils down to modifying the table structure by:
 * Rename the `backends` table to `transport_contexts`.
 * Change its primary key from `backend_name` to auto increment integer.
 * Change `events` foreign key to the new pk of `transport_contexts`.
 * `events` no longer have `priority`.