Fix direct boot mode handling (b/116149732). (#363)
* Fix direct boot mode handling (b/116149732).

The problem was that we were checking for direct boot with a call to
`Context#isDeviceProtectedStorage()` which would always return false as we
never requested a "device protected storage" context. The correct way to
detect direct boot mode is via `UserManager#isUserUnlocked()`.

Additionally the change switches data collection configuration
to use "direct boot" safe storage. It is fine from a security
perspective as we don't store any sensitive information in the config.

* Fix device protected storage handling.