Initial refactor of the plugin structure.
The short-term/immediate goal is to introduce a place for projects to
indicate whether javadoc needs to be generated for them, it's done here
via `firebaseLibrary.publishJavadoc`.

The long-term vision is to move all build and release logic be
controlled via the 'firebase-library' plugin instead of it being
injected into individual projects from the root-project.