Fix Apache license fetching. (#512)
My fix in #484 changed the way the Apache license was handled, and all
whitespace was stripped from this popular license. Essentially, the
original code fetched it directly with the URL class while the other
licenses were fetched with Jsoup.

This commit restores this behavior. This has the unfortunate side-effect
of re-introducing boilerplate code, but it's the most cost-effective
solution for now.