Restrict coverage collection to release builds.
Code coverage can fail for local builds, becuase there are multiple
definitions for the same classes (debug and release). On CI, debug is
disabled. This change forces the coverage check to use on release
builds, thus making it compatible with both local and CI use cases.

This fixes the problem presented in #531.