Remove move product flavors from smoke tests. (#557)
We originally introduced the product flavors to introduce selective test
execution (similar to the check-changed and device-check-changed tasks).
However, we're abandoning this in favor of using the smoke tests as
build tests. Therefore, we must always build with all dependencies and
test cases.

This change keeps the `combined' flavor until we can make changes to the
test runner.