Add new config file for smoke tests.
This change adds the new configuration, which will soon be used by the
runner. This imcreases flexibility by allowing GitHub changes to
immediately affect how the runner operates on CI.

Additionally, this change tweaks the build file to allow testing against
release builds. This will be live once the new runner is deployed.