Increase Remote Config's default timeout.
The default timeout of 5 seconds is way too short. The network is not
guaranteed to be in a good state, and this may cause apps to crash
unexpectedly. This is the source of some flakiness in tests as well.

This change increases the timeout to 30 seconds. Updates are infrequent
and run in the background, so this should not be a problem.