Make component discovery directBootAware (#669)
Firebase itself does not initialize in direct-boot mode and only does so
when the device is unlocked.

While this works most of the time, there are apps that initializeApp()
manually when in direct boot mode and this causes component discovery to
fail(b/138578076).

The change marks ComponentDiscoveryService as directBootAware so that
components are able to be discovered. Whether components are able to
function in direct boot mode is up to each SDK implementation and is
currently provided as best effort.