Only advance the limbo free snapshot if query is synced
Despite what we talked about offline, I still think we need to only advance the last limbo free snapshot version if the query is CURRENT. This is because we only calculate Limbo documents when we are in sync in the server. While we don't persist resume tokens for queries that aren't CURRENT, we do persist the remaining query data (including the lastLimboFreeSnapshot). We should not do that unless we are CURRENT.