Add ServerValue.increment() methods.
* New (currently) package protected method for serverside increments
* Updated ServerValues.resolveFoo to take an existing snapshot
* Added integration tests for new behavior

NOTE: I think I need help undertsanding why I see an integration failure. If my new
tests run then other tests (In EventTest an KeepSyncedTest) fail. It's not clear if
this is because I temporarily go offline or if it's because I've filled the server
queue with an operation the database doesn't yet support and the client cannot handle
the error.