Initial implementation of reflective encoder.
The change introduces the concept of a "fallback" encoder that is used
for any type that does not have an explicitly registered encoder.

The reflective encoder itself is currently implemented in terms of
public getters and does not look at class fields, this can be made
configurable in the future.