Get the IID token before fetching Remote Config. (#1100)
Remote Config requires an IID token to evaluate certain conditions, but does not currently wait for the token from `FirebaseInstanceId#getToken()` before fetching. This returns the default condition value in some cases (ex. percentage conditions) rather than the evaluated result for an app instance.

This change replaces the use of the deprecated `#getToken()` with `#getInstanceId()`, which returns an async Task we can wait to complete before fetching, without otherwise blocking.