Crashlytics Analytics Connector realtime fixes
Refactor the Breadcrumbs receiver to listen for app exception events,
which enables us to block app shutdown until we've confirmed that
Firebase Analytics has received the app exception event.

Because Firebase Analytics serializes events asynchronously, it was possible
(and in fact quite common) for the defaultUncaughtExceptionHandler to close
the app BEFORE Firebase Analytics was finished writing the app exception event
to disk, which ultimately caused the event to be lost. Now Crashlytics waits
for the FA listener to fire with an app exception event before shutting down.

The 2000ms timeout was chosen somewhat arbitrarily; in practice the callback
is received within 100ms or so, and it does not block at all if Firebase
Analytics is not present.