Catching NPE and throwing IOE instead.
Youtube observed this error on some devices when rolling out a new version. b/150526702
We were not able to reproduce this Exception, but found implementations of (Http)URLConnection that might return NULL for #getOutputStream.
The new code is preventing NPE and following the default error scenario that a connection to the FIS server API can not be established.