Add large payload support to datatransport.
Large payloads are stored in a separate table as 80KB chunks of data keyed
by (sequence_num,event_id).
The motivation for the change is to avoid the Android SQLite
implementation limitation, that uses `CursorWindow`s of a limited size
that is unknown at runtime and vendors can change arbitrarily, hence the
conservative choice of 80KB instead of 2048KB that is present in AOSP
source code.

The way it manifests is that it is possible to write rows in the the db
that is impossible to read back as an exception gets thrown.

An altertative approach that was considered was to store file paths in
the DB and store blobs in the file-system. But that would introduce more
complexity to the storage system:

* Loss of ACID semantics, i.e. if we fail to commit we may fail to
  delete the underlying file.
* Should we store it in cache dir or in file dir?
* What if cache is cleared but the event reference still exists in the db?
* etc.