Increase number of retries before deleting events
The current retry logic will delete an event after 10 retries,
which will only take ~17 minutes. If a device is on a bad network
for that period of time, we could relatively easily lose events.

Increasing this to 16 will extend this timeframe to ~18 hours,
giving more opportunity for a retry to occur when the device has
a better network connection.