Call `setDoOutput(true)` on auth POSTs
The FirebaseInstallationsServiceClient wasn't calling `setDoOutput(true)` on the `generateAuthToken` method, which was confusing some SSL providers and causing requests to unceremoniously fail. There's a separate issue causing the task failure not to get correctly propagated back to the caller (in this case FCM), but for now this is an easy and correct-seeming fix.