Disable fetching of Firebase Installation Id when data collection is disabled (#2610)
* Disable fetching of Firebase Installation Id when data collection is disabled

Querying for the Firebase Installation Id makes a network call and potentially registers
new information to Firebase servers, so we should not access it from Crashlytics when
the data collection flag is disabled. Instead, we cache a synthetic FID to use as long
as collection is disabled. This can result in over-counting unique users who have experienced
a crash, but it better aligns with the privacy goals of the data collection flag.

See internal doc go/crashlytics-fid for more details.