Fix Crashlytics user id serialization bug (#3160)
This fixes a bug caught in post-commit validation of PR #3143. The
bug never shipped.

Before the earlier refactor, the user id was serialized in two different
files on disk: one as a raw id and the other in JSON. In the refactor,
this was inadvertedly consolidated to one file path. The result is that
when generating a Crashlytics report, the SDK expects a plain text user id
in the file, and instead sends up the entire JSON blob as the user id.

A follow-on refactor is already underway which will elliminate the need
for the redundant file. Until that change is ready, we'll return to
using distinct file paths for the two files.