Crashlytics efficient keys (#3254)
Refactors UserMetadata and related classes to gain some significant efficiency improvements when calling `setCustomKey`:
* Checks if the map contents have actually changed before writing the key/value store to disk.
* UserMetadata now maintains a "dirty" marker bit for the keys, so calls to serialize the key will no-op when the bit is false.
* Allow at most one queued serialization task at a time, and the queued task will always grab the latest key data when it executes. This prevents a large amount of stale Task objects being created and queued when `setCustomKeys` is invoked repeatedly. (Which has been known to cause OOMs in some extreme cases.)

Also finishes cleanup of extra files from PR #3160: That PR was a quick fix for misreading the user id, which was stored in two different places on disk. This is a proper fix that uses a single source of truth for the serialized user id.