Do not throw when FileStore cannot create a dir. (#3541)
There is a problem with the file structure on some devices #3269. When we did the FileStore refactor, we turned a silent failure into a runtime exception, causing apps to crash on some of these devices. Although this does not fix the root problem, it reverts to the silent behaviour we had before. We still need to investigate the root cause of this.