Deflake firebase_common HeartBeat tests.
The tests relied on `TestOnCompleteListener` that was not safe to call
more than once since it was based on a count down latch. So reusing it
multiple times would cause await() to return immediately.

This change makes it so that a new latch is created for every await()
call, making all await() calls work.

Fixes: http://b/245956774