Change concurrency group for health metrics test to "github.sha" (#4428)
The problem with "github.ref" was that when two consecutive
commits are pushed to a feature branch (e.g. `master`) around
the same time, the execution of health metrics test for the
second commit will cancel the execution for the first one, and
thus make the metrics measurement for the first commit unavailable.