Migrate all sdks to use testFixtures
This overcomes the problem of having multiple testutil source roots
which Android Studio does not recognize. While auto completion still
does not work properly in Android Studio Chipmonk, it will likely start
working in subsequent studio releases. Note that autocomplete does not
work with the testutil approach either, so technically it's not a
regression.