Persist the aqs session id on new aqs session (#5274)
Persist the aqs session id on new aqs session, instead of on event. This
is needed for ANRs and native crashes that can't persist on event.

This also populates the aqs session id and fid in the Report.
The new `CrashlyticsAppQualitySessionsStore` class is based on the
existing `MetaDataStore` class but with the current session id not being
final.