firebase-perf: Add custom method to set final static method for test
Robolectric's ReflectionHelpers removes supporting of setting fnial
static method becasuse it not the reposibility of Robolectric, and
it will not be allowed by the latest JDK. This CL just adds custom
method to do the same things as a workaround. When the project updates
the JDK someday, the maintainer needs to rewrite the test or provides
some test helper method to avoid set final static method for test.

Signed-off-by: utzcoz <utzcoz@outlook.com>