Restructure library group logic in buildSrc (#5297)
Library groups are now compute at the plugin, rather than being a
singleton class using registrars. This fixes the previous bug that
caused duplicate entries to be generated for the release generator
because the singleton class was kept alive by the gradle daemon.

Additionally, it simplifies the verification of missing libraries due to
library groups. Now, the build **will fail** if the release.json file
does not include all the libraries in the library group.