Remove ndk dependencies for test app (#5341)
Remove NDK dependency for test-app
Due to this change:
https://github.com/firebase/firebase-android-sdk/pull/4311/files

There is a step when Acore builds the NDK artifact where they rename
a.library. It needs to be done because Android will ignore any libraries
in the wrong format. When you do a project level dependency, that step
never happens. So remove the NDK dependency and the NDK test cases will
be test on an external app.