Add useReleasedVersions property to sessions e2e test app (#5359)
Add `useReleasedVersions` property to the sessions e2e test app. Also
updated dep versions, and configured test lab to be more consistent.
This is so the e2e test app can run on released versions in the
scheduled workflow, but default to using the project level deps during
local development.