Create post release plugin workflow (#5356)
Per [b/301961341](https://b.corp.google.com/issues/301961341),

This creates a workflow that can be invoked to automatically create a
mergeback branch *and* run the `postReleaseCleanup` task for after a
release.

The workflow is ran from the release branch, but the PR is created for
`master`. I will update the release manual to reflect this as well after
it's merged and I've tested it.

The `name` input is purely used for naming the mergeback branch PR and
such.