Avoid changing api surface (#5406)
Avoid changing the api surface. Having `val`s at the file level causes
Kotlin to generate a public class with the filenameKt which adds a
public class to the public api surface: "The public api surface has
changed for the subproject firebase-sessions: error: Added class
com.google.firebase.sessions.SessionDatastoreKt [AddedClass]". So we
have to put them in a class or companion object.