Use the latest release bom in the sessions test app (#5509)
Use the latest release bom in the sessions test app instead of trying to
figure out released versions.

Also removed direct references to Sessions subscribers from the test app
because it will make the e2e test workflow fail after this feature
merges to main line but before the release. After the release, we can
put that back in.