Make data store pref configs multi-process safe (#5494)
Adds the process name to the config name. This is so two processes don't
try to access the same config on the same file at the same time,
possibly corrupting data. Data store preferences 1.1.0 will support
multi-process. So when that becomes stable, we can upgrade to it and
remove this util. Then all processes will share the same preferences.
This will mostly only affect settings since it will be fetched on each
process, while the sessions store will only happen on the service. But I
made that safe too.