Remove process-details for now. (#5518)
We identified a bug where the backend protobuf is not in agreement with
the structure of the message being sent. This causes the entire message
to be dropped. Making this work will require a bit of refactoring.

Revert this for now to be safe and add it back in a subsequent PR.

Co-authored-by: jrothfeder <rothbutter@google.com>