Implement release notes toggle (#5715)
Per [b/324410387](https://b.corp.google.com/issues/324410387),

This implements a new property on `FirebaseLibrary` for toggling the
generation of release notes. Previously, we were using the presence of
Javadocs to determine this- as libraries without Javadocs don't publish
release notes... at least that's what we thought. We have since realized
that we have multiple libraries that publish release notes, but do *not*
publish Javadocs. To accommodate these libraries, this new property can
be toggled to disable release note generation.

This PR also adds appdistro to its librarygroup, as it seems that is the
intended behavior.

NO_RELEASE_CHANGE