Cleanup Uploader tests (#5793)
Per [b/330185088](https://b.corp.google.com/issues/330185088),

This adds a cleanup step to our uploader tests. This hasn't been needed
up until now, as tests have not only *not* cared about the order of
events- but they make use of all the events they store. With the
addition of our new tests, this is no longer the case. Depending on the
order these new tests run it, it can produce a pass or fail for certain
tests that are ascertaining the order of the events.

As such, the store will now be wiped after each test.