Fix compile type transformations (#5832)
Per [b/277605778](https://b.corp.google.com/issues/277605778),

This removes the step during our release generation where we were
converting the scope of dependencies to compile. As such, all published
libraries that had dependencies on firebase library were changed to
`api` level dependencies. While this may not be necessary for a lot of
them, it's a sufficient middle ground between all dependencies being
`api` level and none. It should now be easier to deduce what exactly
needs to be `api` level, and which does not.