Try catch for datastore write (#5833)
Proposed fix for issue #5802. Datastore crashes when writing from a
device that has full internal memory.

I did a manual test by creating a clean project and importing the part
of the code with the condition checkers to simulate the issue. I was
able to receive the issue stacktrace the developer was encountering and
was able to verify that the try catch condition mitigates the issue.