Added keep rule for Crashlytics Build ID resource names (#5882)
Proposed fix for #5859, previously regressed issue from #5562.

The issue affects both `firebase-crashlytics` and
`firebase-crashlytics-ndk. This behavior is triggered when
`isShrinkResources` is enabled and when "strict" rule. is added to
resource shrinkage. e.g.
```
<?xml version="1.0" encoding="utf-8"?>
<resources xmlns:tools="http://schemas.android.com/tools"
    tools:shrinkMode="strict" />
```

b/311837855 advises that if strict mode is used, the keep rule is needed
for getIdentifier constants.

Simulated tests by publishing local copy of the `firebase-crashlytics`,
and `firebase-crashlytics-ndk`. Then testing with sample app variant
that includes `isShrinkResources` enabled in `build.gradle`

```
buildTypes {
    release {
            isShrinkResources = true
            isMinifyEnabled = true
    }
}
```