Fix copyRootGoogleServices task (#5895)
Per [b/322356188](https://b.corp.google.com/issues/322356188),

This fixes the issues with the `copyRootGoogleServices` task. 

Namely, it seems there are various tasks in the pipeline that have
_implicit_ dependencies on source files- which gradle doesn't like. To
fix this, we just ensure `copyRootGoogleServices` is set as a dependency
of them all.

This PR also moves the `copyRootGoogleServices` task into our android
library plugin- to keep everything together.