Comment the equals() and hashCode() methods in GeneratedConnector and GeneratedConnector
From b/340910059: In order to document the behavior of the equals() and hashCode() methods of the generated classes, these methods were added to the GeneratedConnector and GeneratedOperation interfaces in the Android Core SDK; however, adding these methods is a breaking change that requires the implementing classes to explicitly define them, rather than inheriting the (correct) default behavior from kotlin.Any. Since the released codegen does not explicitly define these methods it fails to compile. However, once the codegen that does define those methods is sufficiently soaked, they should be uncommented.