Improve cold initialization time (#5986)
Improve cold initialization time by not validating the fid on init. We
only need to detect if the fid has changed on initialization, we don't
care if it's valid. Changing it, deleting it, manipulating it,
whatever... should cause the Crashlytics id to rotate. That will still
happen after this change.

In my local experiments, this improved the cold init time from about 300
ms to 50 ms. This does not affect warm init, when the fid and auth token
are already cached. On the first report, if the auth token has not yet
been fetched, it will get fetched at upload time. If validating the fid
caused the fid to rotate, the Crashlytics id will also rotate on the
next launch.