Change default port of useEmulator() to 9399 (was 9510).
In cl/636395279 the grpc and http ports were combined. Also, that CL changes the --grpc_port and --http_port flags into a single flag: --listen, whose default value is 127.0.0.1:3628. This change will be included in the v1.1.18 release of the Data Connect emulator, and will be used by firestore-tools with https://github.com/firebase/firebase-tools/pull/7211 merged in.