Refactor ComponentProvider (#6037)
The ComponentProvider is lightweight dependency injection for Firestore
components. This PR expands the number of class that are instantiated as
part of the ComponentProvider. This separation of concerns simplifies
classes, and makes testing these classes in isolation possible.

This PR includes some refactoring of tests, with respect to
instantiating classes.

This PR prepares for more thorough integration tests as part of cache
clearing logic.