Implement Spotless formatting (#6100)
NO_RELEASE_CHANGELOG

This may require a lot of PRs to rebase/merge. As a resolution,
outstanding PRs could add Spotless themselves like done in the first
commit and run `spotlessApply` before the merge. Perhaps adding Spotless
should be its own commit to main so outstanding branches can just pull
main and cherry pick in?

Updated root documentation, sessions and vertex both still link to ktfmt
for their formatting, how do we want to update that?

Google Java Format was updated to latest. `ktfmt` was *not*. We are
currently using a version from 2022, `0.41`. Latest is `0.51`, however,
upgrading will require a reformat. Do we want to pre-emptively upgrade
now and format with the rest of these formatting changes? Current latest
release of Spotless does not have configuration for ktfmt trailing comma
formatting, which was added recently. It's committed to the repo 3 weeks
ago but they haven't had a release.