Generalize the CrashlyticsWorker (#6126)
Generalize the CrashlyticsWorker.

The old CrashlyticsBackgroundWorker has a lot of overhead. Local
experimentation shows this generalized worker manages submitting tasks
about 4 times as fast.

The old background worker has some inaccurate Javadoc. I corrected and
verified every detail mentioned in the Javadoc with integration tests.
We can rely on it now.

I also found a bug in the old worker. If you cancel a task then submit a
runnable or callable, that will never execute because the chain breaks.
I fixed it in the general worker and added test cases.

Next steps will be to utilize the generic worker for the proposed
common, network, and disk write workers.