Improvements to the CrashlyticsWorker (#6143)
Added a method to submit a task followed by a continuation to the
worker. This is useful for making a continuation happen right after the
submitted task, even if other tasks were submitted to the worker in the
meantime. This will be useful for fetching settings from a network
executor, then continuing on the common worker to parse the settings
json.

Also added a method to race two tasks on the worker. This is useful for
things like waiting for an for explicit data collection enable call, or
automatic data collection being enabled.

Both methods have their behaviour fully documented in the javadoc and
tested in unit tests.