Add a CrashlyticsWorkers container to manage the workers (#6178)
Add a CrashlyticsWorkers container to manage the workers, and define
what each one is for. Also added a `submitTaskOnSuccess` method for
convenience. This will be useful for any of the user actions that get
triggered by send unsent reports.

Also found and fixed a bug in the worker queue. If a task gets cancelled
and is followed by a runnable or callable already in the queue, it would
have been skipped. Now it's not a problem, and a test case added.