Make the session id not depend on fetching the fid (#6152)
Make generating a session id not depend on fetching the fid. This is
needed so we can allow the common worker to initialize, and start
accepting use actions, before any data collection, settings, or fid
validation happens.

The hash of the crashlytics installation id was used to pad the length
and increase entropy. Instead of that, we use a random uuid now.
Verified nothing depends on the source of that padding by manually
setting it to all 0s and doing an end-to-end test.

The way the session id is built is a little bit weird, using bytes. In
the future, we could rewrite this to use proper string formatting.