Switch `RequestOption` to accept long timeout only
The `RequestOptions` class only customizable paramter, `timeout`, is
now called `timeoutInMillis` and it's type is `Long`.

Using `kotlin.time.Duration` in our public API, while convenient,
introduces an issue with Java code. In short, `Duration` is exposed as
`long` in the JVM, and that long can represent time in milliseconds, or
nanoseconds.

For clarity, we've fallback to long.

This change also removes the internal representation of
`RequestOptions`, which was a legacy implementation detail.