Switch `RequestOption` to accept long timeout only (#6289)
The `RequestOptions` class only customizable paramter, `timeout`, is now
called `timeoutInMillis` and it's type is `Long`.

Using `kotlin.time.Duration` in our public API, while convenient,
introduces an issue with Java code. In short, `Duration` is exposed as
`long` in the JVM, and that long can represent time in milliseconds, or
nanoseconds.

For clarity, we've fallback to long.