Fix inefficiency in the setCustomKeys Kotlin extension (#6536)
Fixed inefficiency in the Kotlin `FirebaseCrashlytics.setCustomKeys`
extension.

The old implementation called `setCustomKey` for every key/value pair,
which serializes to disk, and is not an intuitive behaviour for
customers. This now calls `setCustomKeys` once with the entire built
map.

This proxies to the existing Java `CustomKeysAndValues.Builder` but
gives nice Kotlin syntax. The old class had a public ctor, I plan to
make this internal since customers shouldn't instantiate this class
directly. This is done in a way to not break the api, even though the
api is internal.

pair=tejasd