Add clean command to fireci (#6563)
Per [b/381270432](https://b.corp.google.com/issues/381270432),

This adds a command to our `fireci` cli tool that can work as an
alternative to `gradle clean`. There seems to be some dependency issues
preventing `gradle clean` from completely correctly, but with the usage
of `fireci`- we can avoid working within gradle entirely.

Furthermore, this command comes with flags for deeper cleaning of
gradle's caches.

I've also updated the dependencies that fireci was using, as I ran into
a variety of issues trying to use it "as-is" on an M1 machine. The
updated dependencies seem to include fixes for such issues.

Command usage:
```sh
fireci clean --help
```