Refactor Vendor Plugin (#6640)
Per [b/391351857](https://b.corp.google.com/issues/391351857),

This refactors the `VendorPlugin` and `VendorTask` to be more compliant
with the gradle cache and configuration avoidance. It also adds proper
task logging and documentation to both the task and the plugin.

This PR also fixes the following:

- [b/391352175](https://b.corp.google.com/issues/391352175) -> Rename
plugin Util files to Extensions