Implement GMavenService (#6644)
Per [b/392134866](https://b.corp.google.com/issues/392134866),

This implements a centralized interface for communicating with GMaven
called `GMavenService`. This service implements the gradle build-service
interface to provide proper parallel access, and keeps local
`ConcurrentHashMap` instances to cache responses.

Cached responses are on a _per-build_ basis, to avoid improper caching
of dynamic data. That is, given any build- **all** tasks that utilize
`GMavenService` will share the responses from GMaven; even within a
parallel environment. But if the tasks are considered out-of-date and
are ran again, then new requests will be made to the GMaven backend.

Tests and documentation are provided for everything added as well.

Note that while this PR _implements_ `GMavenService`- it does _not_
refactor the existing `GMavenHelper` and `RepositoryClient` usages to
use it. That will occur in subsequent PRs, as to avoid polluting this
PR. Furthermore, while there are no tests for `PomElement` directly- in
a future PR that includes tests for bom generation, `PomElement` will be
tested as a by-product.

This PR also fixes the following:

- [b/392135224](https://b.corp.google.com/issues/392135224) -> Implement
centralized pom datamodel