Bump well known types (#6716)
Per [b/398840288](https://b.corp.google.com/issues/398840288),

This bumps `protolite-well-known-types` to properly utilize `3.25.5`. It
seems as though this was an oversight in #6343, but since gradle uses
the highest version when resolving dependency conflicts (and all the
existing libraries already use `3.25.5`), this isn't a major issue. This
is only really an issue if someone is using `protolite-well-known-types`
in isolation (which isn't really a use-case we're shipping for). But the
main reason for fixing this is that it causes a bit of confusion when
trying to track dependency issues (see issue #6674 for an example of
this).

Fixes #6674