Avoid Process.myProcessName() on Android 13 (#6720)
Avoid calling `Process.myProcessName()` on Android 13 because it appears
to be missing from some OEM-specific Android 13 builds. It is fine to
just let the method fall through to the next, older, method to get the
process name. See
https://github.com/firebase/firebase-unity-sdk/issues/1059

I have not been able to reproduce this issue locally, but this change is
very safe.

We should consider refactoring Crashlytics to consume the Sessions
`ProcessDetails` data class, instead of the current `@AutoValue` holder.